cambios

+++ src/app/pages/teachers/classroom-list/classroom-list.component.ts
@@ -1,11 +1,13 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Classroom } from 'src/api/models/classroom.model';
 import { Teacher } from 'src/api/models/teacher.model';
 import { User } from 'src/api/models/user.model';
 import { ClassroomsService } from 'src/api/services/classrooms-service/classrooms.service';
 import { TeachersService } from 'src/api/services/teachers-service/teachers.service';
 import { UsersService } from 'src/api/services/users-service/users.service';
+import { DeleteComponent } from 'src/app/shared/dialog/dialogs/delete/delete.component';
 
 @Component({
     selector: 'app-student-classroom-list',
@@ -15,12 +17,14 @@ import { UsersService } from 'src/api/services/users-service/users.service';
 export class ClassroomListComponent implements OnInit {
 
     public classrooms: IClassroomList[];
+    public x : boolean;
 
     constructor(
         private classroomsService: ClassroomsService,
         private userService: UsersService,
         private teacherService: TeachersService,
-        private http: HttpClient
+        private http: HttpClient,
+        public dialog: MatDialog
     ) {
 
     }
@@ -47,10 +51,21 @@ export class ClassroomListComponent implements OnInit {
     }
 
     public deleteClass(classroomId: number) {
-        this.classroomsService.deleteClassroom(classroomId)
-            .subscribe(_ => {
-                this.classrooms.splice(this.classrooms.findIndex(c => c.id === classroomId), 1);
-            });
+        console.log("public deleteClass(classroomId: number) {");
+        console.log(classroomId);
+
+
+        let dialogRef = this.dialog.open(DeleteComponent);
+
+        dialogRef.afterClosed().subscribe(result =>{
+            console.log('The dialog was closed')
+            if(result){
+                this.classroomsService.deleteClassroom(classroomId)
+                    .subscribe(_ => {
+                        this.classrooms.splice(this.classrooms.findIndex(c => c.id === classroomId), 1);
+                    });
+            }
+        });
     }
 }
